refactor: move impl of expectation combinator for `Not` to new module `expectation_combinators`
